feat: move credential helper to private function
