+++ pkg/apis/application/v1alpha1/types_test.go
@@ -3881,58 +3881,3 @@ func TestApplicationSpec_GetSourcePtrByIndex(t *testing.T) {
 		})
 	}
 }
-
-func TestRepository_GetCredentialType(t *testing.T) {
-	tests := []struct {
-		name string
-		repo *Repository
-		want string
-	}{
-		{
-			name: "Empty Repository",
-			repo: nil,
-			want: "",
-		},
-		{
-			name: "HTTPS Repository",
-			repo: &Repository{
-				Repo:     "foo",
-				Password: "some-password",
-			},
-			want: "https",
-		},
-		{
-			name: "SSH Repository",
-			repo: &Repository{
-				Repo:          "foo",
-				SSHPrivateKey: "some-key",
-			},
-			want: "ssh",
-		},
-		{
-			name: "GitHub App Repository",
-			repo: &Repository{
-				Repo:                    "foo",
-				GithubAppPrivateKey:     "some-key",
-				GithubAppId:             1,
-				GithubAppInstallationId: 1,
-			},
-			want: "github-app",
-		},
-		{
-			name: "Google Cloud Repository",
-			repo: &Repository{
-				Repo:                 "foo",
-				GCPServiceAccountKey: "some-key",
-			},
-			want: "cloud-source-repositories",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := tt.repo.GetCredentialType(); got != tt.want {
-				t.Errorf("Repository.GetCredentialType() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
